+++ R/utilities-quantity.R
@@ -28,7 +28,6 @@
 #'
 #' # Returns all `Volume` quantities defined in `Organism` and all its subcontainers
 #' quantities <- getAllQuantitiesMatching("Organism|**|Volume", sim)
-#'
 #' @export
 getAllQuantitiesMatching <- function(paths, container) {
   getAllEntitiesMatching(paths, container, Quantity)
@@ -94,18 +93,18 @@ setQuantityValues <- function(quantities, values, units = NULL) {
   values <- c(values)
 
   # Test for correct inputs
-  ospsuite.utils::validateIsOfType(quantities, Quantity)
-  ospsuite.utils::validateIsNumeric(values)
+  validateIsOfType(quantities, Quantity)
+  validateIsNumeric(values)
 
   if (length(values) > 1) {
-    ospsuite.utils::validateIsSameLength(quantities, values)
+    validateIsSameLength(quantities, values)
   } else {
     values <- rep(values, length(quantities))
   }
 
   if (!is.null(units)) {
-    ospsuite.utils::validateIsSameLength(quantities, units)
-    ospsuite.utils::validateIsString(units)
+    validateIsSameLength(quantities, units)
+    validateIsString(units)
   }
 
   for (i in seq_along(quantities)) {
@@ -141,14 +140,14 @@ setQuantityValues <- function(quantities, values, units = NULL) {
 #' setParameterValuesByPath(list("Organism|Liver|Volume", "Organism|Liver|A"), c(2, 3), sim)
 #' @export
 setQuantityValuesByPath <- function(quantityPaths, values, simulation, units = NULL, stopIfNotFound = TRUE) {
-  ospsuite.utils::validateIsString(quantityPaths)
-  ospsuite.utils::validateIsNumeric(values)
-  ospsuite.utils::validateIsSameLength(quantityPaths, values)
-  ospsuite.utils::validateIsOfType(simulation, Simulation)
+  validateIsString(quantityPaths)
+  validateIsNumeric(values)
+  validateIsSameLength(quantityPaths, values)
+  validateIsOfType(simulation, Simulation)
 
   if (!is.null(units)) {
-    ospsuite.utils::validateIsSameLength(quantityPaths, units)
-    ospsuite.utils::validateIsString(units)
+    validateIsSameLength(quantityPaths, units)
+    validateIsString(units)
   }
 
   task <- getContainerTask()
@@ -180,8 +179,8 @@ scaleQuantityValues <- function(quantities, factor) {
   quantities <- c(quantities)
 
   # Test for correct inputs
-  ospsuite.utils::validateIsOfType(quantities, Quantity)
-  ospsuite.utils::validateIsNumeric(factor)
+  validateIsOfType(quantities, Quantity)
+  validateIsNumeric(factor)
 
   lapply(quantities, function(q) q$value <- q$value * factor)
   invisible()
@@ -196,8 +195,8 @@ scaleQuantityValues <- function(quantities, factor) {
 #' @return a display path for each entry in paths
 #'
 getQuantityDisplayPaths <- function(paths, simulation) {
-  ospsuite.utils::validateIsString(paths)
-  ospsuite.utils::validateIsOfType(simulation, Simulation)
+  validateIsString(paths)
+  validateIsOfType(simulation, Simulation)
   displayResolver <- getNetTask("FullPathDisplayResolver")
   paths <- c(paths)
 
@@ -211,4 +210,26 @@ getQuantityDisplayPaths <- function(paths, simulation) {
   })
 
   return(unlist(displayPaths, use.names = FALSE))
+}
+
+
+#' Retrieves the path of all observers defined in the container and all its children
+#'
+#' @param container A Container or Simulation used to find the observers
+#' @seealso [loadSimulation()], [getContainer()] and [getAllContainersMatching()] to retrieve objects of type Container or Simulation
+#'
+#' @return An array with one entry per observer defined in the container
+#' @examples
+#'
+#' simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
+#' sim <- loadSimulation(simPath)
+#'
+#' # Returns the path of all quantities defined in the simulation
+#' observerPaths <- getAllObserverPathsIn(sim)
+#' @export
+getAllObserverPathsIn <- function(container) {
+  return(setdiff(
+    x = getAllQuantityPathsIn(container),
+    y = c(getAllParameterPathsIn(container), getAllMoleculePathsIn(container))
+  ))
 }