+++ tests/dev/script-data-set.R
@@ -4,7 +4,7 @@ library(ospsuite)
 obsData <- loadDataRepositoryFromPKML("tests/data/obs_data.pkml")
 
 # Create an empty data set
-dataSet <- DataSet$new(obsData)
+dataSet <- DataSet$new(dataRepository = obsData)
 
 # get the values
 xValues <- dataSet$xValues
@@ -12,7 +12,7 @@ xValues <- dataSet$xValues
 xUnit <- dataSet$xUnit
 
 
-emptyDataSet <- DataSet$new()
+emptyDataSet <- DataSet$new(name = "NewDataSet")
 
 yVal <- emptyDataSet$yValues
 
@@ -20,6 +20,6 @@ yVal <- emptyDataSet$yValues
 emptyDataSet$xValues <- c(1, 2, 3, 4, 5)
 emptyDataSet$xDimension <- ospDimensions$Amount
 
-dataSet <- DataSet$new()
+dataSet <- DataSet$new(name = "NewDataSet")
 dataSet$addMetaData("Meta", "Value")
 expect_equal(dataSet$metaData[["Meta"]], "Value")